// src/model/profile.ts
export class ProfileViewModel{
    FirstName: string;
    MiddleName: string;
    LastName: string;
    FullName: string;
    Address1: string;
    Address2: string;
    City: string;
    State: string;
    Zip: string;
    Address: string;
    MobileNumber: string;
    Gender: string;
    BirthDate: string;
    Email: string;
    IsAdmin: boolean;
}